Module Hardcaml_xilinx

module Byte_write_width : sig ... end

module Cascade_height : sig ... end

Height of BRAM cascades chains. This can be explicitly specified to help vivado meet timing when it is unnecessarily cascading BRAMs.

module Collision_mode : sig ... end

module Ram_arch : sig ... end

Xilinx RAM primitive types.

module Dual_port_ram : sig ... end

Single clock Dual Port Memory

module Fifo_async : sig ... end

module Fifo_sync : sig ... end

module Simple_dual_port_ram : sig ... end

Simple Dual Port Memory. 1 port is for writing, the other for reading.

module True_dual_port_ram : sig ... end

True Dual Port Memory with independent clocks for ports a and b.

module Ram_port : sig ... end

module Ram_port_with_clear : sig ... end

Statemachine for clearing a RAM via one of it's ports.

module Synthesis : sig ... end

module Memory_builder : sig ... end

A general-purpose means of representing memories. The Config.t type allows the user to configure the underlying memory implementations. Eg: Using URAM for bits 0-72, and BRAMs for bits 73-80. This module allows construction of memories in 1D or 2D Modes. See further documentation below.